+++ src/main/java/roomescape/domain/reservation/repository/impl/ThemeDAO.java
@@ -117,7 +117,6 @@ public class ThemeDAO implements ThemeRepository {
 
     @Override
     public List<Theme> findPopularThemes(LocalDate startDate, LocalDate endDate) {
-//        T.id, T.name, T.description, T.thumbnail
         String sql = """
                 select T.*, count(R.id) as reservation_count
                 from theme T
