+++ src/test/java/roomescape/domain/reservation/controller/ThemeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.controller;
+package roomescape.domain.reservation.controller;
 
 public class ThemeControllerTest {
     // TODO 작성
