--- src/test/java/roomescape/reservation/repository/impl/ThemeDAOTest.java
@@ -1,14 +1,0 @@
-package roomescape.reservation.repository.impl;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-class ThemeDAOTest {
-
-    @DisplayName("인기있는 테마를 반환한다.")
-    @Test
-    void test1(){
-
-    }
-
-}
